[KMSDRM/ATOMIC] Move the atomic interface functions to their own file.
